refactor: extract BundleDiscovery helper
Move the seven-tier bundle discovery out of cmd/purge-quarantine.rb
into Homebrew::CaskTools::BundleDiscovery
(lib/cask_tools/bundle_discovery.rb) so external consumers — the
babble external command (master plan W3) — can reuse discovery
without the quarantine removal. Pure refactor; the tier logic,
messages, cache path, and TTL are unchanged.

Public API: .new(token, cask_dir), #bundles (the tier chain),
#candidate_names (public, memoized), .lsregister_dump (class-level;
shared on-disk cache). The command keeps run, the xattr operations,
and all quiet-gating, and delegates discovery.

Notes:

- Namespace is Homebrew::CaskTools, NOT the Homebrew::Cask sketched
  in the master plan: defining Homebrew::Cask shadows ::Cask for
  brew code written inside module Homebrew (lexical constant
  resolution precedes the Object fallback), breaking e.g.
  cleanup.rb and diagnostic.rb at runtime.
- Tier-transition --verbose gating moves from args.verbose? to
  Context (brew.rb sets Context.current from parsed args, so CLI
  behavior is identical; Utils::Output::Mixin already uses the
  same fallback).
- run-tests.sh and ci.yml hardlink lib/ and test/lib/ into
  $(brew --repo): the hardlinked command resolves
  require_relative against the hardlink location.
- Discovery specs move to test/lib/cask_tools/bundle_discovery_spec.rb
  (brew tests --only=lib/cask_tools/bundle_discovery); the command
  spec keeps xattr coverage and gains a delegation example.
