+++ .github/workflows/ci.yml
@@ -78,9 +78,12 @@ jobs:
       - name: Link tap files into Homebrew for brew tests
         run: |
           brew_lib="$(brew --repo)/Library/Homebrew"
-          mkdir -p "${brew_lib}/cmd" "${brew_lib}/test/cmd"
+          mkdir -p "${brew_lib}/cmd" "${brew_lib}/test/cmd" \
+                   "${brew_lib}/lib/cask_tools" "${brew_lib}/test/lib/cask_tools"
           ln -f cmd/purge-quarantine.rb "${brew_lib}/cmd/purge-quarantine.rb"
           ln -f test/cmd/purge-quarantine_spec.rb "${brew_lib}/test/cmd/purge-quarantine_spec.rb"
+          ln -f lib/cask_tools/bundle_discovery.rb "${brew_lib}/lib/cask_tools/bundle_discovery.rb"
+          ln -f test/lib/cask_tools/bundle_discovery_spec.rb "${brew_lib}/test/lib/cask_tools/bundle_discovery_spec.rb"
           ln -f cmd/cask-extract.rb "${brew_lib}/cmd/cask-extract.rb"
           ln -f test/cmd/cask-extract_spec.rb "${brew_lib}/test/cmd/cask-extract_spec.rb"
           ln -f cmd/man.rb "${brew_lib}/cmd/man.rb"
@@ -91,6 +94,7 @@ jobs:
       - name: brew tests
         run: |
           brew tests --only=cmd/purge-quarantine
+          brew tests --only=lib/cask_tools/bundle_discovery
           brew tests --only=cmd/cask-extract
           brew tests --only=cmd/man
           brew tests --only=cmd/generate-tap-man-completions
@@ -101,6 +105,10 @@ jobs:
           brew_lib="$(brew --repo)/Library/Homebrew"
           rm -f "${brew_lib}/cmd/purge-quarantine.rb"
           rm -f "${brew_lib}/test/cmd/purge-quarantine_spec.rb"
+          rm -f "${brew_lib}/lib/cask_tools/bundle_discovery.rb"
+          rm -f "${brew_lib}/test/lib/cask_tools/bundle_discovery_spec.rb"
+          rmdir "${brew_lib}/lib/cask_tools" "${brew_lib}/lib" \
+                "${brew_lib}/test/lib/cask_tools" "${brew_lib}/test/lib" 2>/dev/null || true
           rm -f "${brew_lib}/cmd/cask-extract.rb"
           rm -f "${brew_lib}/test/cmd/cask-extract_spec.rb"
           rm -f "${brew_lib}/cmd/man.rb"
